feat: Refactor enemy spawning and movement system
- Introduced modular classes for enemy management, including EnemyFactory, EnemyMovement, and WaveSpawnManager.
- Enhanced enemy creation logic with dynamic spawning and special enemy types.
- Improved enemy movement handling with collision detection and behavior adjustments.
- Added spawn position management for better enemy placement.
- Updated existing functions for backward compatibility and streamlined enemy-related logic.
- Implemented performance tracking for dynamic difficulty adjustments during gameplay.
